refactor(outbound): migrate infrastructure persistence, caching, and PKI to outbound adapters
