--- src/adapters/memory.rs
@@ -1,716 +1,0 @@
-//! In-memory adapters for application-service unit tests and memory-only use.
-#[cfg(any(
-    feature = "server",
-    feature = "postgres",
-    feature = "sqlite",
-    feature = "mysql"
-))]
-use crate::domain::StatusListSnapshot;
-#[cfg(any(
-    feature = "server",
-    feature = "postgres",
-    feature = "sqlite",
-    feature = "mysql"
-))]
-use crate::ports::StatusListHistoryRepository;
-use crate::{
-    domain::{Credential, StatusListRecord},
-    ports::{CredentialRepository, PortError, StatusListCache, StatusListRepository},
-};
-use async_trait::async_trait;
-use std::{collections::HashMap, sync::Arc};
-use tokio::sync::RwLock;
-
-#[derive(Clone, Default)]
-pub struct MemoryStatusLists {
-    values: Arc<RwLock<HashMap<String, StatusListRecord>>>,
-    // Shared snapshot storage so `update_with_snapshot` can write the row and
-    // its snapshot under one held lock — the memory mirror of the SQL adapter's
-    // single transaction. `None` when history is disabled (memory-only builds do
-    // not compile the history port at all).
-    #[cfg(any(
-        feature = "server",
-        feature = "postgres",
-        feature = "sqlite",
-        feature = "mysql"
-    ))]
-    history: Option<Arc<RwLock<HashMap<String, StatusListSnapshot>>>>,
-}
-
-#[cfg(any(
-    feature = "server",
-    feature = "postgres",
-    feature = "sqlite",
-    feature = "mysql"
-))]
-impl MemoryStatusLists {
-    /// Shares `history`'s storage so the atomic write paths can record the row
-    /// and its snapshot together; without it the repo would drop snapshots.
-    pub fn with_history(mut self, history: &MemoryStatusListHistory) -> Self {
-        self.history = Some(history.values.clone());
-        self
-    }
-
-    /// Borrows the shared snapshot storage, refusing to proceed when it was
-    /// never wired up.
-    ///
-    /// The atomic write paths promise the row and its snapshot land together.
-    /// Silently skipping the snapshot while still reporting success would make
-    /// this double certify the exact split the port forbids — and a test
-    /// asserting "the snapshot was persisted" would pass vacuously against a
-    /// `MemoryStatusLists::default()`. Failing here surfaces the missing
-    /// `.with_history(..)` as a loud wiring error instead.
-    #[cfg(any(
-        feature = "server",
-        feature = "postgres",
-        feature = "sqlite",
-        feature = "mysql"
-    ))]
-    fn require_history(
-        &self,
-    ) -> Result<&Arc<RwLock<HashMap<String, StatusListSnapshot>>>, PortError> {
-        self.history
-            .as_ref()
-            .ok_or_else(|| PortError::StorageUnavailable {
-                operation: crate::ports::PortOperation::UpdateStatusList,
-                detail: "MemoryStatusLists was built without shared history storage; \
-                         construct it with `.with_history(..)` so snapshots are \
-                         persisted atomically with the row write"
-                    .to_string(),
-            })
-    }
-}
-
-#[async_trait]
-impl StatusListRepository for MemoryStatusLists {
-    async fn find(&self, id: &str) -> Result<Option<Arc<StatusListRecord>>, PortError> {
-        Ok(self.values.read().await.get(id).cloned().map(Arc::new))
-    }
-    async fn insert(&self, record: StatusListRecord) -> Result<(), PortError> {
-        self.values
-            .write()
-            .await
-            .insert(record.list_id.clone(), record);
-        Ok(())
-    }
-    /// Mirrors the SQL adapter's optimistic guard so concurrency behavior is
-    /// testable without a database: the write lands only if the stored stamp is
-    /// still `expected_updated_at`.
-    async fn update(
-        &self,
-        record: StatusListRecord,
-        expected_updated_at: i64,
-    ) -> Result<bool, PortError> {
-        let mut values = self.values.write().await;
-        match values.get(&record.list_id) {
-            Some(current) if current.updated_at == expected_updated_at => {}
-            _ => return Ok(false),
-        }
-        values.insert(record.list_id.clone(), record);
-        Ok(true)
-    }
-    /// Atomic mirror of the SQL adapter's transactional update: the row CAS and
-    /// snapshot insert both happen under the row-map write lock, so no reader
-    /// sees the row advanced without its snapshot. Lock order is always row-map
-    /// then history-map (never the reverse), so this cannot deadlock.
-    #[cfg(any(
-        feature = "server",
-        feature = "postgres",
-        feature = "sqlite",
-        feature = "mysql"
-    ))]
-    async fn update_with_snapshot(
-        &self,
-        record: StatusListRecord,
-        expected_updated_at: i64,
-        snapshot: StatusListSnapshot,
-    ) -> Result<bool, PortError> {
-        let history = self.require_history()?;
-        let mut values = self.values.write().await;
-        match values.get(&record.list_id) {
-            Some(current) if current.updated_at == expected_updated_at => {}
-            _ => return Ok(false),
-        }
-        history
-            .write()
-            .await
-            .insert(snapshot.snapshot_id.clone(), snapshot);
-        values.insert(record.list_id.clone(), record);
-        Ok(true)
-    }
-    /// Atomic mirror of the SQL adapter's transactional insert: the row and the
-    /// snapshot covering its initial state both land under the row-map write
-    /// lock, so no reader sees a published list without its snapshot. Same lock
-    /// order as `update_with_snapshot` (row-map then history-map).
-    #[cfg(any(
-        feature = "server",
-        feature = "postgres",
-        feature = "sqlite",
-        feature = "mysql"
-    ))]
-    async fn insert_with_snapshot(
-        &self,
-        record: StatusListRecord,
-        snapshot: StatusListSnapshot,
-    ) -> Result<(), PortError> {
-        let history = self.require_history()?;
-        let mut values = self.values.write().await;
-        history
-            .write()
-            .await
-            .insert(snapshot.snapshot_id.clone(), snapshot);
-        values.insert(record.list_id.clone(), record);
-        Ok(())
-    }
-    /// Mirrors the SQL adapter's `GROUP BY sub ORDER BY sub` via a `BTreeSet`
-    /// (dedup + sort), so this double doesn't diverge from production semantics.
-    async fn list_uris(&self) -> Result<Vec<String>, PortError> {
-        let uris: std::collections::BTreeSet<String> = self
-            .values
-            .read()
-            .await
-            .values()
-            .map(|r| r.sub.clone())
-            .collect();
-        Ok(uris.into_iter().collect())
-    }
-}
-#[derive(Clone, Default)]
-pub struct MemoryStatusListCache {
-    values: Arc<RwLock<HashMap<String, StatusListRecord>>>,
-}
-#[async_trait]
-impl StatusListCache for MemoryStatusListCache {
-    async fn get(&self, id: &str) -> Result<Option<Arc<StatusListRecord>>, PortError> {
-        Ok(self.values.read().await.get(id).cloned().map(Arc::new))
-    }
-    async fn put(&self, record: StatusListRecord) -> Result<(), PortError> {
-        self.values
-            .write()
-            .await
-            .insert(record.list_id.clone(), record);
-        Ok(())
-    }
-    async fn invalidate(&self, id: &str) -> Result<(), PortError> {
-        self.values.write().await.remove(id);
-        Ok(())
-    }
-}
-
-#[derive(Clone, Default)]
-pub struct MemoryCredentials {
-    values: Arc<RwLock<HashMap<String, Credential>>>,
-}
-
-#[async_trait]
-impl CredentialRepository for MemoryCredentials {
-    async fn find(&self, issuer: &str) -> Result<Option<Credential>, PortError> {
-        Ok(self.values.read().await.get(issuer).cloned())
-    }
-
-    async fn insert(&self, credential: Credential) -> Result<(), PortError> {
-        self.values
-            .write()
-            .await
-            .insert(credential.issuer.0.clone(), credential);
-        Ok(())
-    }
-}
-
-#[cfg(any(
-    feature = "server",
-    feature = "postgres",
-    feature = "sqlite",
-    feature = "mysql"
-))]
-#[derive(Clone, Default)]
-pub struct MemoryStatusListHistory {
-    values: Arc<RwLock<HashMap<String, StatusListSnapshot>>>,
-}
-
-#[cfg(any(
-    feature = "server",
-    feature = "postgres",
-    feature = "sqlite",
-    feature = "mysql"
-))]
-#[async_trait]
-impl StatusListHistoryRepository for MemoryStatusListHistory {
-    async fn insert(&self, record: StatusListSnapshot) -> Result<(), PortError> {
-        self.values
-            .write()
-            .await
-            .insert(record.snapshot_id.clone(), record);
-        Ok(())
-    }
-
-    async fn find_valid_at(
-        &self,
-        list_id: &str,
-        time: i64,
-    ) -> Result<Option<StatusListSnapshot>, PortError> {
-        let values = self.values.read().await;
-        let result = values
-            .values()
-            .filter(|r| r.list_id == list_id && r.iat <= time && r.exp > time)
-            .max_by_key(|r| r.iat)
-            .cloned();
-        Ok(result)
-    }
-
-    async fn delete_older_than(&self, cutoff: i64) -> Result<u64, PortError> {
-        let mut values = self.values.write().await;
-        let to_remove: Vec<String> = values
-            .values()
-            .filter(|r| r.exp < cutoff)
-            .map(|r| r.snapshot_id.clone())
-            .collect();
-        let count = to_remove.len() as u64;
-        for id in to_remove {
-            values.remove(&id);
-        }
-        Ok(count)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{
-        application::{GetStatusListToken, PublishStatusList, UpdateStatuses, UseCaseError},
-        domain::{DomainError, Issuer, Status, StatusEntry, StatusList, StatusListRecord},
-        ports::StatusListCache,
-    };
-    fn record() -> StatusListRecord {
-        StatusListRecord {
-            list_id: "id".into(),
-            issuer: Issuer("issuer".into()),
-            status_list: StatusList {
-                bits: 1,
-                lst: "".into(),
-            },
-            sub: "https://example/id".into(),
-            updated_at: 1672531200,
-        }
-    }
-
-    #[tokio::test]
-    async fn application_services_work_without_infrastructure() {
-        let repo = Arc::new(MemoryStatusLists::default());
-        let cache = Arc::new(MemoryStatusListCache::default());
-        PublishStatusList::new(repo.clone())
-            .execute(record())
-            .await
-            .unwrap();
-        assert!(matches!(
-            PublishStatusList::new(repo.clone()).execute(record()).await,
-            Err(UseCaseError::AlreadyExists)
-        ));
-        let fetched = GetStatusListToken::new(repo.clone(), cache.clone())
-            .execute("id")
-            .await
-            .unwrap();
-        assert_eq!(fetched.list_id, "id");
-        UpdateStatuses::new(repo, cache.clone())
-            .execute(&Issuer("issuer".into()), "id", Vec::new())
-            .await
-            .unwrap();
-        assert!(cache.get("id").await.unwrap().is_none());
-    }
-
-    #[cfg(any(
-        feature = "server",
-        feature = "postgres",
-        feature = "sqlite",
-        feature = "mysql"
-    ))]
-    #[tokio::test]
-    async fn application_services_work_with_history() {
-        use crate::application::{PublishStatusListWithHistory, UpdateStatusesWithHistory};
-        let cache = Arc::new(MemoryStatusListCache::default());
-        let history = Arc::new(MemoryStatusListHistory::default());
-        // Share the history storage so the atomic update path can record its
-        // snapshot (mirrors the SQL adapter, where both writes hit one db).
-        let repo = Arc::new(MemoryStatusLists::default().with_history(history.as_ref()));
-        let token_exp_secs = 900u64;
-        PublishStatusListWithHistory::new(repo.clone(), history.clone(), token_exp_secs)
-            .execute(record())
-            .await
-            .unwrap();
-        assert!(matches!(
-            PublishStatusListWithHistory::new(repo.clone(), history.clone(), token_exp_secs)
-                .execute(record())
-                .await,
-            Err(UseCaseError::AlreadyExists)
-        ));
-        let fetched = GetStatusListToken::new(repo.clone(), cache.clone())
-            .execute("id")
-            .await
-            .unwrap();
-        assert_eq!(fetched.list_id, "id");
-        UpdateStatusesWithHistory::new(repo.clone(), cache.clone(), history.clone())
-            .execute(&Issuer("issuer".into()), "id", Vec::new(), token_exp_secs)
-            .await
-            .unwrap();
-        assert!(cache.get("id").await.unwrap().is_none());
-        // The atomic update committed its snapshot: the row's new stamp resolves
-        // to a stored historical snapshot.
-        let updated_at = repo.find("id").await.unwrap().unwrap().updated_at;
-        assert!(
-            history
-                .find_valid_at("id", updated_at)
-                .await
-                .unwrap()
-                .is_some(),
-            "the atomic update must persist a history snapshot for the new stamp"
-        );
-    }
-
-    #[tokio::test]
-    async fn publish_use_case_enforces_serialized_list_size_limit() {
-        // The size limit must live in the use case, not the handler: an inbound
-        // adapter reaching straight for `PublishStatusList` (bypassing the web
-        // layer) must still be unable to persist an over-limit list.
-        let repo = Arc::new(MemoryStatusLists::default());
-        let mut oversized = record();
-        oversized.status_list.lst = "serialized-list-well-over-the-limit".to_string();
-        let result = PublishStatusList::new(repo.clone())
-            .with_max_serialized_list_size(1)
-            .execute(oversized)
-            .await;
-        assert!(matches!(result, Err(UseCaseError::StatusListTooLarge)));
-        // Nothing was persisted: the guard rejects before insert.
-        assert!(repo.find("id").await.unwrap().is_none());
-    }
-
-    #[tokio::test]
-    async fn update_statuses_rejects_wrong_issuer() {
-        let repo = Arc::new(MemoryStatusLists::default());
-        let cache = Arc::new(MemoryStatusListCache::default());
-        PublishStatusList::new(repo.clone())
-            .execute(record())
-            .await
-            .unwrap();
-
-        let result = UpdateStatuses::new(repo, cache)
-            .execute(&Issuer("other-issuer".into()), "id", Vec::new())
-            .await;
-
-        assert!(matches!(result, Err(UseCaseError::IssuerMismatch)));
-    }
-
-    #[tokio::test]
-    async fn get_status_list_reports_not_found() {
-        let repo = Arc::new(MemoryStatusLists::default());
-        let cache = Arc::new(MemoryStatusListCache::default());
-
-        let result = GetStatusListToken::new(repo, cache)
-            .execute("missing")
-            .await;
-
-        assert!(matches!(result, Err(UseCaseError::NotFound)));
-    }
-
-    #[tokio::test]
-    async fn update_statuses_propagates_domain_errors() {
-        let repo = Arc::new(MemoryStatusLists::default());
-        let cache = Arc::new(MemoryStatusListCache::default());
-        PublishStatusList::new(repo.clone())
-            .execute(record())
-            .await
-            .unwrap();
-
-        let result = UpdateStatuses::new(repo, cache)
-            .execute(
-                &Issuer("issuer".into()),
-                "id",
-                vec![StatusEntry {
-                    index: 0,
-                    status: Status::ApplicationSpecific(3),
-                }],
-            )
-            .await;
-
-        assert!(matches!(
-            result,
-            Err(UseCaseError::Domain(DomainError::InvalidStatusList(_)))
-        ));
-    }
-
-    /// Serves a stale `updated_at` from `find` while delegating everything
-    /// else, modeling a racing writer that advanced the stamp between this
-    /// writer's read and its guarded write.
-    struct StaleReadStatusLists {
-        inner: MemoryStatusLists,
-        stale_updated_at: i64,
-    }
-
-    #[async_trait]
-    impl StatusListRepository for StaleReadStatusLists {
-        async fn find(&self, id: &str) -> Result<Option<Arc<StatusListRecord>>, PortError> {
-            Ok(self.inner.find(id).await?.map(|record| {
-                let mut stale = record.as_ref().clone();
-                stale.updated_at = self.stale_updated_at;
-                Arc::new(stale)
-            }))
-        }
-        async fn insert(&self, record: StatusListRecord) -> Result<(), PortError> {
-            self.inner.insert(record).await
-        }
-        async fn update(
-            &self,
-            record: StatusListRecord,
-            expected_updated_at: i64,
-        ) -> Result<bool, PortError> {
-            self.inner.update(record, expected_updated_at).await
-        }
-        #[cfg(any(
-            feature = "server",
-            feature = "postgres",
-            feature = "sqlite",
-            feature = "mysql"
-        ))]
-        async fn update_with_snapshot(
-            &self,
-            record: StatusListRecord,
-            expected_updated_at: i64,
-            snapshot: StatusListSnapshot,
-        ) -> Result<bool, PortError> {
-            self.inner
-                .update_with_snapshot(record, expected_updated_at, snapshot)
-                .await
-        }
-        #[cfg(any(
-            feature = "server",
-            feature = "postgres",
-            feature = "sqlite",
-            feature = "mysql"
-        ))]
-        async fn insert_with_snapshot(
-            &self,
-            record: StatusListRecord,
-            snapshot: StatusListSnapshot,
-        ) -> Result<(), PortError> {
-            self.inner.insert_with_snapshot(record, snapshot).await
-        }
-        async fn list_uris(&self) -> Result<Vec<String>, PortError> {
-            self.inner.list_uris().await
-        }
-    }
-
-    /// The memory adapter's CAS must reproduce the SQL guard semantics: a
-    /// write guarded on a stale stamp is rejected as `Conflict`, and the
-    /// rejected write leaves no trace — the cache entry survives untouched.
-    #[tokio::test]
-    async fn update_statuses_returns_conflict_when_write_races() {
-        let inner = MemoryStatusLists::default();
-        inner.insert(record()).await.unwrap();
-        let repo = Arc::new(StaleReadStatusLists {
-            inner,
-            stale_updated_at: 100,
-        });
-        let cache = Arc::new(MemoryStatusListCache::default());
-        cache.put(record()).await.unwrap();
-
-        let result = UpdateStatuses::new(repo, cache.clone())
-            .execute(&Issuer("issuer".into()), "id", Vec::new())
-            .await;
-
-        assert!(matches!(result, Err(UseCaseError::Conflict)));
-        assert!(
-            cache.get("id").await.unwrap().is_some(),
-            "a rejected write must not invalidate the cache"
-        );
-    }
-
-    /// Same race through the history-aware use case: nothing may be persisted
-    /// for a write that never landed — no snapshot, no cache invalidation.
-    #[cfg(any(
-        feature = "server",
-        feature = "postgres",
-        feature = "sqlite",
-        feature = "mysql"
-    ))]
-    #[tokio::test]
-    async fn update_statuses_with_history_writes_nothing_on_conflict() {
-        use crate::application::UpdateStatusesWithHistory;
-        let history = Arc::new(MemoryStatusListHistory::default());
-        // Wire the shared history so the repo *could* record a snapshot: the
-        // assertion that it records nothing on conflict is then meaningful.
-        let inner = MemoryStatusLists::default().with_history(history.as_ref());
-        inner.insert(record()).await.unwrap();
-        let repo = Arc::new(StaleReadStatusLists {
-            inner,
-            stale_updated_at: 100,
-        });
-        let cache = Arc::new(MemoryStatusListCache::default());
-        cache.put(record()).await.unwrap();
-        let token_exp_secs = 900u64;
-
-        let result = UpdateStatusesWithHistory::new(repo, cache.clone(), history.clone())
-            .execute(&Issuer("issuer".into()), "id", Vec::new(), token_exp_secs)
-            .await;
-
-        assert!(matches!(result, Err(UseCaseError::Conflict)));
-        assert!(cache.get("id").await.unwrap().is_some());
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap()
-            .as_secs() as i64;
-        assert!(
-            history.find_valid_at("id", now).await.unwrap().is_none(),
-            "a rejected write must not record a historical snapshot"
-        );
-    }
-
-    /// A repo built without shared history storage must refuse the atomic write
-    /// paths rather than dropping the snapshot and reporting success. Without
-    /// this, any future test asserting "the snapshot was persisted" would pass
-    /// vacuously against a `default()` repo, and the double would certify the
-    /// very row-without-snapshot split the port forbids.
-    #[cfg(any(
-        feature = "server",
-        feature = "postgres",
-        feature = "sqlite",
-        feature = "mysql"
-    ))]
-    #[tokio::test]
-    async fn atomic_writes_fail_loudly_without_shared_history() {
-        use crate::domain::{StatusList as DomainStatusList, StatusListSnapshot};
-
-        let repo = MemoryStatusLists::default(); // deliberately not `.with_history(..)`
-        repo.insert(record()).await.unwrap();
-
-        let snapshot = StatusListSnapshot {
-            snapshot_id: "snap".into(),
-            list_id: "id".into(),
-            issuer: Issuer("issuer".into()),
-            status_list: DomainStatusList {
-                bits: 1,
-                lst: "".into(),
-            },
-            sub: "https://example/id".into(),
-            iat: 1672531200,
-            exp: 1672532100,
-        };
-
-        let updated = repo
-            .update_with_snapshot(record(), 1672531200, snapshot.clone())
-            .await;
-        assert!(
-            updated.is_err(),
-            "update_with_snapshot must not report success while dropping the snapshot"
-        );
-
-        let inserted = repo.insert_with_snapshot(record(), snapshot).await;
-        assert!(
-            inserted.is_err(),
-            "insert_with_snapshot must not report success while dropping the snapshot"
-        );
-    }
-
-    /// The publish path records the snapshot covering the list's initial state
-    /// through the repository, so a §8.4 lookup at the publish instant resolves
-    /// immediately rather than falling into a permanent hole.
-    #[cfg(any(
-        feature = "server",
-        feature = "postgres",
-        feature = "sqlite",
-        feature = "mysql"
-    ))]
-    #[tokio::test]
-    async fn publish_with_history_records_the_initial_snapshot() {
-        use crate::application::PublishStatusListWithHistory;
-        let history = Arc::new(MemoryStatusListHistory::default());
-        let repo = Arc::new(MemoryStatusLists::default().with_history(history.as_ref()));
-
-        PublishStatusListWithHistory::new(repo.clone(), history.clone(), 900)
-            .execute(record())
-            .await
-            .unwrap();
-
-        let published = repo.find("id").await.unwrap().unwrap();
-        let snapshot = history
-            .find_valid_at("id", published.updated_at)
-            .await
-            .unwrap()
-            .expect("publish must record a snapshot covering the initial state");
-        assert_eq!(snapshot.iat, published.updated_at);
-        assert_eq!(snapshot.list_id, "id");
-    }
-
-    /// End-to-end pin of the stamp monotonicity through the use case and the
-    /// memory adapter: two updates inside the same wall-clock second must both
-    /// advance `updated_at`, or the second writer's guard could not move.
-    #[tokio::test]
-    async fn update_statuses_advances_updated_at_within_same_second() {
-        let repo = Arc::new(MemoryStatusLists::default());
-        let cache = Arc::new(MemoryStatusListCache::default());
-        PublishStatusList::new(repo.clone())
-            .execute(record())
-            .await
-            .unwrap();
-        let initial = repo.find("id").await.unwrap().unwrap().updated_at;
-
-        let update = UpdateStatuses::new(repo.clone(), cache);
-        update
-            .execute(&Issuer("issuer".into()), "id", Vec::new())
-            .await
-            .unwrap();
-        let first = repo.find("id").await.unwrap().unwrap().updated_at;
-        update
-            .execute(&Issuer("issuer".into()), "id", Vec::new())
-            .await
-            .unwrap();
-        let second = repo.find("id").await.unwrap().unwrap().updated_at;
-
-        assert!(first > initial, "update must advance the stamp");
-        assert!(
-            second > first,
-            "a same-second update must still advance the stamp"
-        );
-    }
-
-    /// Pins the read-through cache's consistency model (C2). `GetStatusListToken`
-    /// serves whatever the cache holds without revalidating against the
-    /// repository, so a value cached before a repository change stays visible
-    /// until it is invalidated (or its TTL lapses). This is the deliberately
-    /// bounded staleness window — documented and pinned here so nobody later
-    /// assumes the getter provides strong read-after-write consistency. The fix
-    /// (if one is ever wanted) is version-tagged puts; that is out of scope for a
-    /// bounded, inherited window.
-    #[tokio::test]
-    async fn get_status_list_serves_stale_cache_until_invalidated() {
-        let repo = Arc::new(MemoryStatusLists::default());
-        let cache = Arc::new(MemoryStatusListCache::default());
-        repo.insert(record()).await.unwrap();
-
-        // A reader populates the cache with v1.
-        let getter = GetStatusListToken::new(repo.clone(), cache.clone());
-        assert_eq!(
-            getter.execute("id").await.unwrap().sub,
-            "https://example/id"
-        );
-
-        // The repository advances to v2 out of band, without touching the cache
-        // (models a writer whose invalidation has not landed, or another process).
-        let mut v2 = record();
-        v2.updated_at += 1;
-        v2.sub = "https://example/v2".into();
-        repo.insert(v2).await.unwrap();
-
-        // The getter still serves the cached v1: read-through does not revalidate.
-        assert_eq!(
-            getter.execute("id").await.unwrap().sub,
-            "https://example/id"
-        );
-
-        // Only invalidation (what UpdateStatuses does after a write) closes the
-        // window; the next read then repopulates from the repository.
-        cache.invalidate("id").await.unwrap();
-        assert_eq!(
-            getter.execute("id").await.unwrap().sub,
-            "https://example/v2"
-        );
-    }
-}
