--- src/adapters/mod.rs
@@ -1,16 +1,0 @@
-#[cfg(feature = "aws")]
-pub mod aws;
-#[cfg(feature = "cache-moka")]
-pub mod cache;
-#[cfg(feature = "certificate-acme")]
-pub mod certificate;
-pub mod memory;
-#[cfg(feature = "redis")]
-pub mod redis;
-#[cfg(any(
-    feature = "server",
-    feature = "postgres",
-    feature = "sqlite",
-    feature = "mysql"
-))]
-pub mod sea_orm;
