+++ src/outbound/mod.rs
@@ -1,0 +1,6 @@
+pub mod aws;
+pub mod cache;
+pub mod cert;
+pub mod memory;
+pub mod redis;
+pub mod sql;
